requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers==2.2.2
pydantic==1.10.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
